+++ index.ios.js
@@ -1,0 +1,6 @@
+
+var PlayInstallReferrer = {};
+
+PlayInstallReferrer.getInstallReferrerInfo = function() {}
+
+module.exports = { PlayInstallReferrer }
